Rename blob_log_reader.{h,cc} to blob_log_sequential_reader.{h,cc}
